Added: - add to process map - delete from process map
Registering all the running Actors in the new Supervisor namespace ETProcesses

Rewrote the .Act() function, and added one separate for root so we can add all the processes to the ETProcesses

Added a Process.Stop() function that will cancel the ETFunc, and also delete the process from the procMap.

Fixed missing mutex'es when using maps.

Added mutex to the supervisor process map